refactor : 패키징 변경 및 README.md 파일 구현클래스 정의 수정
