+++ src/main/java/racingcar/domain/racing/Message.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.racing;
 
 public enum Message {
     GET_CARS_NAME_MESSAGE("경주 할 자동차 이름(이름은 쉼표(,) 기준으로 구분)"),
