+++ src/main/java/racingcar/domain/racing/RandomNumber.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.racing;
 
 import camp.nextstep.edu.missionutils.Randoms;
