+++ src/main/java/racingcar/domain/result/WinNames.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.result;
 
 import java.util.List;
