+++ src/test/java/racingcar/domain/TurnTest.java
@@ -2,6 +2,7 @@ package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.racing.Turn;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
