+++ src/test/java/racingcar/domain/WinnerTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racingcar.domain.car.CarName;
+import racingcar.domain.result.Result;
+import racingcar.domain.result.WinNames;
 
 import java.util.*;
